requirements.txt
# Core ML
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.36.0
accelerate>=0.25.0
peft>=0.7.0
bitsandbytes>=0.41.0
datasets>=2.15.0
sentencepiece>=0.1.99

# Vision
Pillow>=10.0.0
opencv-python-headless>=4.8.0

# LLM APIs
openai>=1.6.0

# Vector search (for optional similarity component)
faiss-cpu>=1.7.4

# Data processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Logging & monitoring
loguru>=0.7.0
tqdm>=4.65.0
wandb>=0.16.0

# Evaluation & visualization
matplotlib>=3.7.0
seaborn>=0.12.0
tabulate>=0.9.0

# Demo
streamlit>=1.28.0
gradio>=4.0.0

# Code quality
black>=23.0.0
ruff>=0.1.0
pytest>=7.4.0
pytest-cov>=4.1.0

# Utilities
requests>=2.31.0
httpx>=0.25.0
tenacity>=8.2.0
rank-bm25>=0.2.2
sentence-transformers>=2.2.0
